+++ sdk/storage/storage-file-share/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Features Added
 
+- Added `ShareFileClient.listRanges()` and `ShareFileClient.listRangesDiff()` to list file ranges as a paginated async iterable with continuation-token support.
+- Deprecated `ShareFileClient.getRangeList()` and `ShareFileClient.getRangeListDiff()` in favor of `listRanges()` and `listRangesDiff()`.
+
 ### Breaking Changes
 
 ### Bugs Fixed